apps/CalcServer/requirements.txt
numpy
pandas
scipy
PyKrige
matplotlib
geopandas
shapely
PyYAML